perf(ordering): MVV-LVA capture ordering and a wider move score
Replace the crude victim-only capture score (piece_at + 1) with proper most-valuable-victim / least-valuable-attacker ordering, and widen ScoredMove::score from int8_t to int32_t so captures/killers/history can be layered (captures and queen promotions rank above all quiets via a CAPTURE_SCORE band). ~11% faster on the fixed-depth tactics suite (139.9s -> 124.0s), same moves found; Debug ctest 31/31.

+++ src/move.h
@@ -98,13 +98,16 @@ public:
 // Moves which contain a score
 struct ScoredMove : public Move
 {
-    int8_t score;
+    // Wide enough to layer capture / killer / history scores (see aBMoveValue and
+    // the move selectors): captures rank above all quiets, killers above the rest
+    // of the quiets, and remaining quiets are ordered by the (signed) history score.
+    int32_t score;
 
     // This is so that we can assign efficiently Move object to a *ScoredMove object
     void operator=(Move m) { data = m.getData(); }
 
     // Constructor to initialize both data (from Move) and score
-    ScoredMove(int dataValue = 0, int8_t scoreValue = 0)
+    ScoredMove(int dataValue = 0, int32_t scoreValue = 0)
         : Move(dataValue), score(scoreValue) {}
 
 };
